fix: create user adjustments

+++ src/main/java/br/ufrn/imd/SIGResAPI/controller/UserController.java
@@ -1,9 +1,13 @@
 package br.ufrn.imd.SIGResAPI.controller;
 
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -14,8 +18,12 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import br.ufrn.imd.SIGResAPI.dto.CreateUserDTO;
 import br.ufrn.imd.SIGResAPI.dto.UserDTO;
+import br.ufrn.imd.SIGResAPI.enums.ERole;
+import br.ufrn.imd.SIGResAPI.models.Role;
 import br.ufrn.imd.SIGResAPI.models.User;
+import br.ufrn.imd.SIGResAPI.repository.RoleRepository;
 import br.ufrn.imd.SIGResAPI.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -30,19 +38,34 @@ public class UserController {
     @Autowired
     PasswordEncoder passwordEncoder;
 
+    @Autowired
+    RoleRepository roleRepository;
+
     @GetMapping
     public ResponseEntity<List<User>> allUsers() {
         return ResponseEntity.ok(userRepository.findAll());
     }
 
     @PostMapping
-    public ResponseEntity<User> createUser(@RequestBody UserDTO body) {
+    public ResponseEntity<User> createUser(@RequestBody CreateUserDTO body) {
+        // Verifica se o username já existe
         if (userRepository.existsByUsername(body.username())) {
             return ResponseEntity.badRequest().build();
         }
-        User user = new User(null, body.username(), passwordEncoder.encode(body.password()), body.roles(), null, null,
-                null);
+
+        // Converte os nomes das roles para objetos Role
+        Set<Role> roles = new HashSet<>();
+        for (String roleName : body.roles()) {
+            ERole eRole = ERole.valueOf(roleName); // Obtém o enum ERole baseado no nome da role
+            Role role = roleRepository.findByName(eRole)
+                    .orElseThrow(() -> new RuntimeException("Role not found")); // Busca a Role no banco de dados
+            roles.add(role);
+        }
+
+        // Cria o usuário com as roles
+        User user = new User(null, body.username(), passwordEncoder.encode(body.password()), roles, null, null, null);
         userRepository.save(user);
+
         return ResponseEntity.ok(user);
     }
 
@@ -66,9 +89,17 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (user.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         if (!userRepository.existsById(id)) {
             return ResponseEntity.notFound().build();
         }
+
         userRepository.deleteById(id);
         return ResponseEntity.noContent().build();
     }
